TechTreeをコンポーネント化
